# Conflicts:
#	.vs/CaladriusⅡ/v15/Server/sqlite3/storage.ide
#	Assembly-CSharp-Editor.csproj
#	Assembly-CSharp-firstpass.csproj
#	Assembly-CSharp.csproj
#	Assets/Scenes/GamePlay.unity
#	Assets/Scenes/iwatani.GamePlay.unity
#	Assets/Scenes/iwatani.unity
#	Assets/Scripts/BossEnemy.cs
#	Assets/Scripts/Gage3.cs
#	Library/ScriptAssemblies/Assembly-CSharp-firstpass.dll
#	Library/ScriptAssemblies/Assembly-CSharp-firstpass.pdb
#	Library/ScriptAssemblies/Unity.Analytics.DataPrivacy.dll
#	Library/ScriptAssemblies/Unity.Analytics.DataPrivacy.pdb
#	Library/ScriptAssemblies/Unity.CollabProxy.Editor.dll
#	Library/ScriptAssemblies/Unity.CollabProxy.Editor.pdb
#	Library/ScriptAssemblies/Unity.PackageManagerUI.Editor.dll
#	Library/ScriptAssemblies/Unity.TextMeshPro.Editor.dll
#	Library/ScriptAssemblies/Unity.TextMeshPro.Editor.pdb
#	Library/ScriptAssemblies/Unity.TextMeshPro.dll
#	Library/ScriptAssemblies/Unity.TextMeshPro.pdb
#	Library/ScriptAssemblies/Unity.Timeline.Editor.dll
#	Library/ScriptAssemblies/Unity.Timeline.dll
#	Library/metadata/71/710e1f82c61398346b580ddbb2e78293
#	Library/metadata/b0/b08613f915d912246896836141e8c0ff
#	Library/metadata/b2/b20cabe09f4964b4187fae9e50db9eab
#	Library/shadercompiler-UnityShaderCompiler.exe3.log
#	Library/shadercompiler-UnityShaderCompiler.exe4.log
#	Library/shadercompiler-UnityShaderCompiler.exe5.log
#	Library/shadercompiler-UnityShaderCompiler.exe6.log
#	Library/shadercompiler-UnityShaderCompiler.exe7.log

+++ Assets/Scripts/Enemy.cs
@@ -8,12 +8,11 @@ public class Enemy : MonoBehaviour
     public int rotspeed;
     public int power = 1;
 
-    public int hp = 1;
+    public float hp = 1;
 
 
 
     Bullet playerBullet;
-
     PlayerShield playerShield;
 
     // Start is called before the first frame update
@@ -61,6 +60,19 @@ public class Enemy : MonoBehaviour
                 Debug.Log("敵死亡");
             }
         }
+        //if (col.gameObject.tag == "Beem")
+        //{
+        //    playerBullet = col.gameObject.GetComponent<Bullet>();
+
+        //    Debug.Log("敵Hit");
+        //    // ヒットポイントを減らす
+        //    hp = hp - playerBullet.power;
+        //    if (hp <= 0)
+        //    {
+        //        Destroy(gameObject);
+        //        Debug.Log("敵死亡");
+        //    }
+        //}
 
         if (col.gameObject.tag == "PlayerShield")
         {
